updated structure

+++ Lab2_1_BareMetal/lib/SRV/TaskButtonMonitor/TaskButtonMonitor.cpp
@@ -3,40 +3,44 @@
 #include "LedDriver.h"
 #include "Signals.h"
 
-#define PRESS_THRESHOLD_TICKS 25  // 25 * 20ms = 500ms
-#define LED_OFF_TICKS         15  // 15 * 20ms = 300ms
+/* 25 ticks × 20ms = 500ms threshold */
+#define PRESS_THRESHOLD_TICKS 25
+/* 15 ticks × 20ms = 300ms LED on time */
+#define LED_OFF_TICKS         15
 
-static bool wasPressed = false;
+static bool     wasPressed     = false;
 static uint16_t pressTickCount = 0;
-static int16_t ledOffCount = 0;
-static bool greenActive = false;
+static int16_t  ledOffCount    = 0;
+static bool     greenActive    = false; // true=green, false=red
 
 void TaskButtonMonitor_Run(void)
 {
     bool pressed = ButtonDriver_Read();
 
-    // Rising edge: button just pressed
+    /* Rising edge: button just pressed */
     if (pressed && !wasPressed)
     {
         pressTickCount = 0;
     }
 
-    // Count while held
+    /* Count ticks while held */
     if (pressed)
     {
         pressTickCount++;
     }
 
-    // Falling edge: button just released
+    /* Falling edge: button just released */
     if (!pressed && wasPressed)
     {
-        uint16_t duration = pressTickCount * 20; // ms
+        uint16_t duration = (uint16_t)(pressTickCount * 20); // convert ticks → ms
         bool isShort = (pressTickCount < PRESS_THRESHOLD_TICKS);
 
+        /* Publish to Signals for Task2 */
         Signals_SetLastPressDuration(duration);
         Signals_SetIsShortPress(isShort);
         Signals_SetNewPressFlag(true);
 
+        /* Visual feedback */
         if (isShort)
         {
             LedDriver_GreenOn();
@@ -52,7 +56,7 @@ void TaskButtonMonitor_Run(void)
 
     wasPressed = pressed;
 
-    // LED off timer
+    /* Non-blocking LED off timer */
     if (ledOffCount > 0)
     {
         ledOffCount--;
